Use get_project_root() & get_test_contract_path()
