+++ raiden/tests/utils/tests.py
@@ -1,11 +1,23 @@
-# -*- coding: utf8 -*-
+# -*- coding: utf-8 -*-
 import gc
-
 import gevent
+import os
+
+from raiden.utils import get_project_root
 
 __all__ = (
     'cleanup_tasks',
 )
+
+
+def get_test_contract_path(contract_name):
+    contract_path = os.path.join(
+        get_project_root(),
+        'tests',
+        'smart_contracts',
+        contract_name
+    )
+    return os.path.realpath(contract_path)
 
 
 def cleanup_tasks():
